функционал калькулятора вынесен в отдельный модуль функционал калькулятора добавлен в отдельный модуль
